More SLAP applied for main class of the chatBot
* Added helper class 'Container'
