+++ src/main/java/Main.java
@@ -7,7 +7,7 @@ import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
 
 /**
- * A GUI for Duke using FXML.
+ * A GUI for Taylor using FXML.
  */
 public class Main extends Application {
