+++ src/main/java/tasklist/FindTask.java
@@ -3,8 +3,8 @@ package tasklist;
 import java.util.List;
 
 import exceptions.TaylorException;
-import helper.IterateList;
-import helper.WordsSplit;
+import taskhelper.IterateList;
+import taskhelper.WordsSplit;
 import tasks.Task;
 import ui.Ui;
