+++ src/main/java/tasklist/SearchTask.java
@@ -5,9 +5,9 @@ import java.util.ArrayList;
 import java.util.List;
 
 import exceptions.TaylorException;
-import helper.IterateList;
-import helper.TaskInsertion;
-import helper.WordsSplit;
+import taskhelper.IterateList;
+import taskhelper.TaskInsertion;
+import taskhelper.WordsSplit;
 import tasks.Task;
 import ui.Ui;
